intercept new secscan API calls with suppressions
